+++ packages/bits/demo/src/components/demo/range-filter/index.ts
@@ -1,0 +1,4 @@
+export * from "./range-filter-docs/range-filter-docs.example.component";
+export * from "./range-filter-basic/range-filter-basic.example.component";
+export * from "./range-filter-guides/range-filter-guides.example.component";
+export * from "./range-filter-vertical/range-filter-vertical.example.component";