+++ packages/bits/demo/src/components/demo/range-filter/range-filter-docs/range-filter-docs.example.component.ts
@@ -1,0 +1,8 @@
+import { Component } from "@angular/core";
+
+@Component({
+    selector: "nui-range-filter-docs-example",
+    templateUrl: "./range-filter-docs.example.component.html",
+    standalone: false,
+})
+export class RangeFilterDocsExampleComponent {}