+++ packages/bits/src/lib/range-filter/public-api.ts
@@ -1,0 +1,3 @@
+export * from "./range-filter.component";
+export * from "./range-filter.models";
+export * from "./range-filter.module";